requirements.txt
torch==2.1.0
transformers==4.36.0
datasets==2.16.0
tokenizers==0.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
mlflow==2.9.2
wandb==0.16.1
fastapi
uvicorn
pydantic
boto3==1.34.0
sagemaker==2.199.0
pgvector==0.2.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
onnx==1.15.0
onnxruntime==1.16.3
accelerate==0.25.0
evaluate==0.4.1
seqeval==1.2.2